- refactoring log report functions - implement can submit report
