+++ tcp.py
@@ -53,7 +53,9 @@ class CaBotTCP():
         self.app = Flask(__name__)
         self.cabot_manager = cabot_manager
         self.manage_cabot_char = common.CabotManageChar(self, "manage_cabot", cabot_manager)
-        self.log_request_char = common.CabotLogRequestChar(self, "log_request", cabot_manager)
+        self.log_request_char = common.CabotLogRequestChar(self, "log_request",
+                                                            cabot_manager, 
+                                                            common.CabotLogResponseChar(self, "log_response"))
         self.log_char = common.CabotLogChar(self, "log", cabot_manager)
         self.summons_char = common.SummonsChar(self, "summons")
         self.destination_char = common.DestinationChar(self, "destination")
@@ -101,7 +103,7 @@ class CaBotTCP():
 
         self.speak_char = common.SpeakChar(self, "speak")
         self.event_char = common.EventChars(self, "navigate")
-        self.log_response_char = common.CabotLogResponseChar(self, "log_response")
+
         self.handler = subchar_handler("/cabot")
         self.sio.register_namespace(self.handler)
         self.app.wsgi_app = socketio.WSGIApp(self.sio, wsgi_app=self.app.wsgi_app)
@@ -126,11 +128,6 @@ class CaBotTCP():
 
     def handleEventCallback(self, msg, request_id):
         self.event_char.handleEventCallback(msg, request_id)
-
-    def logResponse(self, response):
-        common.logger.info("cabot log response")
-        self.log_response_char.response(response)
-
 
     def start(self):
         common.logger.info("CaBotTCP thread started")
